% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotype.R
\name{add_clonotype_id}
\alias{add_clonotype_id}
\title{add_clonotype_id}
\usage{
add_clonotype_id(x, name = "clonotype_id")
}
\arguments{
\item{x}{a tidy VDJ object}

\item{name}{name of the column to store clonotype id.}
}
\description{
add_clonotype_id
}
