% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotype.R
\name{add_clonotype_size}
\alias{add_clonotype_size}
\title{add_clonotype_size}
\usage{
add_clonotype_size(x, by = "clonotype_id", name = "clonotype_size")
}
\arguments{
\item{x}{a tidy VDJ object}

\item{by}{name of the column containing clonotype ids.}

\item{name}{name of the column to store clonotype size.}
}
\description{
add_clonotype_size
}
