% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotype.R
\name{clonotype_size}
\alias{clonotype_size}
\title{clonotype_size}
\usage{
clonotype_size(x, keep = "barcode")
}
\arguments{
\item{x}{a tidy VDJ object}

\item{keep}{keep this column(s) when reporting clonotype sizes.}
}
\description{
clonotype_size
}
