% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{compute_vdj_qc}
\alias{compute_vdj_qc}
\title{compute_vdj_qc}
\usage{
compute_vdj_qc(
  x,
  barcode.col = "barcode",
  chain.col = "chain",
  max.alpha = 1,
  max.beta = 1,
  drop_multi = TRUE,
  check_valid = TRUE
)
}
\arguments{
\item{x}{a data.frame with VDJ information.}

\item{barcode.col}{which column use as barcode or index.}

\item{chain.col}{which column contians the chain information.}

\item{max.alpha}{maximum number of alpha chains for singletons.}

\item{max.beta}{maximum number of beta chains for singletons.}

\item{drop_multi}{drop chain "Multi" from 10x VDJ data.}

\item{check_valid}{logical; whether to check for validity.}
}
\value{
a data.frame containing QC information for each barcode/chain.
}
\description{
Calculates some qc values for VDJ data including validity (cells must have at least 1 TRA and 1 TRB chain) and
doublet based on the maximum number of TRA and TRB chains (default 1 for both).
}
