% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vdj_usage}
\alias{plot_vdj_usage}
\title{plot_vdj_usage}
\usage{
plot_vdj_usage(
  x,
  cols = NULL,
  split.by = NULL,
  title = NULL,
  subtitle = NULL,
  size = 3,
  alpha = 0.4,
  fill = "violetred",
  fill.vdj = "limegreen",
  axis.width = 0.3,
  expand.x = 0.1,
  expand.y = 5,
  order = FALSE
)
}
\arguments{
\item{x}{data.frame}

\item{cols}{columns to include.}

\item{split.by}{combine and split the vdj plot by this columns.}

\item{title}{main title.}

\item{subtitle}{subtitlte.}

\item{size}{size of text in gene boxes.}

\item{alpha}{alpha of gene boxes.}

\item{fill}{fill color of gene boxes.}

\item{fill.vdj}{fill color of arrows.}

\item{axis.width}{width of gene boxes.}

\item{expand.x}{expand x axis by this amount.}

\item{expand.y}{expand y axis by this amount.}

\item{order}{whether to sort the plots by number of sequences when using split.by.}
}
\description{
Plot V/J gene usage in alpha/beta chains
}
