% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_vdj_10x}
\alias{read_vdj_10x}
\title{read_vdj_10x}
\usage{
read_vdj_10x(filename, productive.only = TRUE)
}
\arguments{
\item{filename}{filename.}

\item{productive.only}{logical; whether to filter out non-productive rows.}
}
\description{
Read a 10x VDJ file.
}
