% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_vdj_10x_batch}
\alias{read_vdj_10x_batch}
\title{read_vdj_10x_batch}
\usage{
read_vdj_10x_batch(path = ".", extension = "csv",
  productive.only = TRUE)
}
\arguments{
\item{path}{path name to the folder.}

\item{extension}{extension of the 10x VDJ files.}

\item{productive.only}{logical; whether to filter out non-productive rows.}
}
\description{
Read a batch of 10x VDJ files from a folder.
}
