% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{read_vdj_json}
\alias{read_vdj_json}
\title{read_vdj_json}
\usage{
read_vdj_json(filename)
}
\arguments{
\item{filename}{path to consensus_annotations.json file.}
}
\description{
Parses consensus_annotations.json from 10x VDJ.
}
