% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_vdjdb_full}
\alias{read_vdjdb_full}
\title{read_vdjdb_full}
\usage{
read_vdjdb_full(filename, min.score = NULL, paired = TRUE)
}
\arguments{
\item{filename}{filename.}

\item{min.score}{if not NULL, return only entries with this score or greater.}

\item{paired}{return only entries with paired (alpha and beta) VJ information.}
}
\description{
read_vdjdb_full
}
