% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{remove_vdj_doublets}
\alias{remove_vdj_doublets}
\title{remove_vdj_doublets}
\usage{
remove_vdj_doublets(x, by = "barcode", cutoff = 1)
}
\arguments{
\item{x}{data.frame.}

\item{by}{column to use as index.}

\item{cutoff}{cutoff for doublets.}
}
\description{
Remove potential VDJ doublets by making sure there is at much 2 gene variants per cell.
}
