% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_vdj_10x}
\alias{tidy_vdj_10x}
\title{tidy_vdj_10x}
\usage{
tidy_vdj_10x(x, by = "cell_index", drop.na = TRUE, drop.multi = TRUE)
}
\arguments{
\item{x}{data.frame with VDJ data.}

\item{by}{column to use as index.}

\item{drop.na}{logical; whether to drop rows with NAs.}

\item{drop.multi}{logical; whether to drop rows with Multi chains.}
}
\description{
Tidy a 10x VDJ format so that each line contains one barcode-alpha-beta-combination.
}
